adding a few finishing touches
added a feature where upon revealing 2 cards, if they dont match, the game will take a 1 second time out in which the cards opened will be revealed and the player will not be able to reveal any more cards until the time out is over.
also added a check on opening the 2nd card, if the cards dont match, the game will enter the timeout (as written above). and then will flip the cards back, the game no longer needs the player to attempt opening a 3rd card to flip the cards back.
upon finishing the game, the game will return to the original state and let the player press the start button to start a new game.